feat(user-support): split support list and ticket creation pages
Move ticket creation form to /soporte/nuevo, keep /soporte as ticket list with a create CTA, and wire the new user support route.
